services/ai-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
langchain==0.1.20
langchain-community==0.0.38
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
supabase==2.4.6
pdfplumber==0.11.0
spacy==3.7.4
httpx==0.27.0
python-dotenv==1.0.1